+++ CHANGELOG.md
@@ -3,6 +3,14 @@ The **goal** of this file is explaining to the users of our project the notable 
 
 _The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)_.
 
+## [0.5.8] - 2019-08-01
+### Added
+- skip user based on profile bio
+
+### Fixed
+- xpath error likers from photo
+
+
 ## [0.5.7] - 2019-07-24
 ### Fixed
 - user agent error in firefox